Make the groups more meaningful
We want the groups to be similar to Arch Linux's use:

- base:       a minimal sane set of packages for a system.
- base-devel: all the packages required to build other packages (eg
              in a chroot).

This aligns our use of groups with how Arch uses them, although our
base group is much simpler. When actually building we may not have
a package for some of the native packages in base-devel so in that
case we will still have to record the dependency explicitly.